Feat: Added a network bound resource file
This utility function consists of type parameters ResultType and RequestType. The first parameter is a query that gets a list of news articles. The second parameter is a fetch that will get new data from the rest api. The third parameter will store any data from fetch to database. The last parameter decides if it's time to get new data from the database or web data if it is fine for now.